Change in updateMap interface
newMID = updateMap(MIDToBeUpdated, newMap, <optional>deletedMap)

newMID is the the physical reference to the newly inserted newMap.
Its possible that the newMID is reused. In the sense that the older
version is deleted from the Mapfile and the newMap is placed in that
location.

deletedMap is an optional parameter to recieve the version that was
removed. This Map could be empty if nothing was replaced.

Example usage:
```
  Map deletedMap = new Map();
  MID newMID = Mapfile.updateMap(MIDToBeUpdated, newMap, deletedMap);
  if (newMID.isReused) {
    deletedMap.getTimeStatamp();
    deletedMap.getValue();
  }
```